+++ pkg/authenticator/requests.go
@@ -20,11 +20,11 @@ func LoginRequest(authnURL string, conjurVersion string, csrBytes []byte, userna
 		authenticateURL = fmt.Sprintf("%s/inject_client_cert", authnURL)
 	}
 
-	log.InfoLogger.Printf(log.CAKC011I, authenticateURL)
+	log.Info(log.CAKC011I, authenticateURL)
 
 	req, err := http.NewRequest("POST", authenticateURL, bytes.NewBuffer(csrBytes))
 	if err != nil {
-		return nil, log.RecordedError(log.CAKC024E, err.Error())
+		return nil, log.RecordedError(log.CAKC024E, err)
 	}
 	req.Header.Set("Content-Type", "text/plain")
 	req.Header.Set("Host-Id-Prefix", usernamePrefix)
@@ -44,10 +44,10 @@ func AuthenticateRequest(authnURL string, conjurVersion string, account string, 
 		authenticateURL = fmt.Sprintf("%s/%s/%s/authenticate", authnURL, account, url.QueryEscape(username))
 	}
 
-	log.InfoLogger.Printf(log.CAKC012I, authenticateURL)
+	log.Info(log.CAKC012I, authenticateURL)
 
 	if req, err = http.NewRequest("POST", authenticateURL, nil); err != nil {
-		return nil, log.RecordedError(log.CAKC023E, err.Error())
+		return nil, log.RecordedError(log.CAKC023E, err)
 	}
 
 	req.Header.Set("Content-Type", "text/plain")
@@ -60,7 +60,7 @@ func readBody(resp *http.Response) ([]byte, error) {
 
 	responseBytes, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		return nil, log.RecordedError(log.CAKC022E, err.Error())
+		return nil, log.RecordedError(log.CAKC022E, err)
 	}
 
 	return responseBytes, err