+++ pkg/utils/file_test.go
@@ -3,6 +3,8 @@ package utils
 import (
 	"errors"
 	"fmt"
+	"io/ioutil"
+	"os"
 	"testing"
 
 	. "github.com/smartystreets/goconvey/convey"
@@ -54,6 +56,29 @@ func TestFile(t *testing.T) {
 				ShouldResemble,
 				expectedOutput,
 			)
+		})
+	})
+
+	Convey("VerifyFileExists", t, func() {
+		Convey("An existing file returns nil", func() {
+			path := "/tmp/test_file"
+			dataStr := "some\ntext\n"
+			err := ioutil.WriteFile(path, []byte(dataStr), 0644)
+			if err != nil {
+				t.FailNow()
+			}
+
+			err = VerifyFileExists(path)
+
+			So(err, ShouldBeNil)
+		})
+
+		Convey("A non-existing file returns the error", func() {
+			path := "non/existing/path"
+
+			err := VerifyFileExists(path)
+
+			So(os.IsNotExist(err), ShouldBeTrue)
 		})
 	})
 }